Rename utils/utils.js to utils/utils.ts and change all references.
utils.js was also removed from .eslintignore since it is expected to lint it correctly
