+++ tools/cli/commands-packages.js
@@ -4,7 +4,7 @@ var files = require('../fs/files');
 var buildmessage = require('../utils/buildmessage.js');
 var auth = require('../meteor-services/auth.js');
 var config = require('../meteor-services/config.js');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var httpHelpers = require('../utils/http-helpers.js');
 var compiler = require('../isobuild/compiler.js');
 var catalog = require('../packaging/catalog/catalog.js');
