+++ tools/cli/main.js
@@ -16,7 +16,7 @@ var catalog = require('../packaging/catalog/catalog.js');
 var buildmessage = require('../utils/buildmessage.js');
 var httpHelpers = require('../utils/http-helpers.js');
 const archinfo = require('../utils/archinfo');
-import { isEmacs } from "../utils/utils.js";
+import { isEmacs } from "../utils/utils";
 
 var main = exports;
