+++ tools/isobuild/build-plugin.js
@@ -2,7 +2,7 @@ var archinfo = require('../utils/archinfo');
 var buildmessage = require('../utils/buildmessage.js');
 var files = require('../fs/files');
 var _ = require('underscore');
-import utils from '../utils/utils.js';
+import utils from '../utils/utils';
 
 let nextId = 1;
