+++ tools/isobuild/meteor-npm.js
@@ -11,7 +11,7 @@ var os = require('os');
 var _ = require('underscore');
 var httpHelpers = require('../utils/http-helpers.js');
 var buildmessage = require('../utils/buildmessage.js');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var runLog = require('../runners/run-log.js');
 var Profile = require('../tool-env/profile').Profile;
 import { parse } from "semver";
