+++ tools/isobuild/package-api.js
@@ -1,7 +1,7 @@
 var assert = require("assert");
 var _ = require("underscore");
 var buildmessage = require('../utils/buildmessage.js');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var compiler = require('./compiler.js');
 var archinfo = require('../utils/archinfo');
 var catalog = require('../packaging/catalog/catalog.js');
