+++ tools/packaging/catalog/catalog-remote.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var sqlite3 = require('sqlite3');
 
 var files = require('../../fs/files');
-var utils = require('../../utils/utils.js');
+var utils = require('../../utils/utils');
 var buildmessage = require('../../utils/buildmessage.js');
 var config = require('../../meteor-services/config.js');
 var archinfo = require('../../utils/archinfo');
