+++ tools/packaging/package-client.js
@@ -4,7 +4,7 @@ var config = require('../meteor-services/config.js');
 var httpHelpers = require('../utils/http-helpers.js');
 var release = require('./release.js');
 var files = require('../fs/files');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var buildmessage = require('../utils/buildmessage.js');
 var compiler = require('../isobuild/compiler.js');
 var authClient = require('../meteor-services/auth-client.js');
