+++ tools/packaging/package-map.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var packageVersionParser = require('./package-version-parser.js');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 import { isIsobuildFeaturePackage } from '../isobuild/compiler.js';
 
 // PackageMap: Represents the choices of package versions being used for a
