+++ tools/packaging/warehouse.js
@@ -42,7 +42,7 @@ var _ = require("underscore");
 var files = require('../fs/files');
 var httpHelpers = require('../utils/http-helpers.js');
 var fiberHelpers = require('../utils/fiber-helpers.js');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 
 // Use `METEOR_WAREHOUSE_URLBASE` to override the default warehouse
 // url base.
