+++ tools/project-context.js
@@ -12,7 +12,7 @@ import { loadIsopackage } from './tool-env/isopackets.js';
 var packageMapModule = require('./packaging/package-map.js');
 var release = require('./packaging/release.js');
 var tropohouse = require('./packaging/tropohouse.js');
-var utils = require('./utils/utils.js');
+var utils = require('./utils/utils');
 var watch = require('./fs/watch');
 var Profile = require('./tool-env/profile').Profile;
 import { KNOWN_ISOBUILD_FEATURE_PACKAGES } from './isobuild/compiler.js';
