+++ tools/runners/run-mongo.js
@@ -1,7 +1,7 @@
 import { MongoExitCodes } from '../utils/mongo-exit-codes';
 
 var files = require('../fs/files');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var fiberHelpers = require('../utils/fiber-helpers.js');
 var runLog = require('./run-log.js');
 var child_process = require('child_process');
