+++ tools/tests/compiler-plugins.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var selftest = require('../tool-testing/selftest.js');
 var files = require('../fs/files');
 import { getUrl } from '../utils/http-helpers.js';
-import { sleepMs } from '../utils/utils.js';
+import { sleepMs } from '../utils/utils';
 import { host } from '../utils/archinfo';
 const osArch = host();
