+++ tools/tests/mongo.js
@@ -1,6 +1,6 @@
 var selftest = require('../tool-testing/selftest.js');
 var Sandbox = selftest.Sandbox;
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var net = require('net');
 var Future = require('fibers/future');
 var _ = require('underscore');
