+++ tools/tests/package-tests.js
@@ -4,7 +4,7 @@ var selftest = require('../tool-testing/selftest.js');
 var Sandbox = selftest.Sandbox;
 var files = require('../fs/files');
 var testUtils = require('../tool-testing/test-utils.js');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var packageClient = require('../packaging/package-client.js');
 var catalog = require('../packaging/catalog/catalog.js');
