+++ tools/tests/wipe-all-packages.js
@@ -1,7 +1,7 @@
 var selftest = require('../tool-testing/selftest.js');
 var Sandbox = selftest.Sandbox;
 var files = require('../fs/files');
-var utils = require('../utils/utils.js');
+var utils = require('../utils/utils');
 var archinfo = require('../utils/archinfo');
 var _ = require('underscore');
