+++ tools/tool-testing/run.js
@@ -20,7 +20,7 @@ import {
 import { Console } from '../console/console.js';
 import Matcher from './matcher.js';
 import OutputLog from './output-log.js';
-import { randomPort, timeoutScaleFactor, sleepMs } from '../utils/utils.js';
+import { randomPort, timeoutScaleFactor, sleepMs } from '../utils/utils';
 import TestFailure from './test-failure.js';
 import { execFileSync } from '../utils/processes';
